Split tests for readme links from readme examples
